fix -C option to check for approved strings count
download.sh has -C option to download files only if 100% translated.
This uses get_file_status function to check if file is “completed”; but
the logic was checking string count to completed string count and if
not all strings approved, this is not right; it should compare approved
strings to completed strings to determine 100% complete.